pulled out super class into seperate document, minor color changes
