+++ src/style.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 body {
-  @apply bg-gray-800;
+  @apply bg-gray-700;
   font-family: Roboto;
 }
 
@@ -14,7 +14,14 @@ body {
 .main-button {
     @apply bg-gray-800 hover:bg-gray-900 text-gray-300 font-bold py-2 px-10 rounded border-gray-600 border rounded-2xl;
     text-transform: uppercase;
-    
+}
+
+.app-button {
+    @apply bg-gray-800 hover:bg-gray-900 text-gray-300 rounded px-2 py-1 border border-gray-600;
+}
+
+.div-reg-flex {
+    @apply flex flex-col h-screen;
 }
 
 .tx-button-qr {
@@ -29,6 +36,10 @@ body {
     @apply text-center px-5 py-3 bg-gray-800;
     box-shadow: rgb(0 0 0 / 3%) 0px 0px 1px, rgb(0 0 0 / 6%) 0px 5px 10px, rgb(0 0 0 / 12%) 0px 20px 36px;
   }
+
+.section-background-lite {
+    @apply text-center px-5 py-3 bg-gray-700;
+}
 
 .connect-wallet-panel {
     @apply text-center px-5 py-3;
